[2732] Move validation service to a custom validator
